# dmestore-ui/src/main/angular/api-doc.json
{
    "swagger": "2.0",
    "info": {
        "description": "SpingCloud web接口列表",
        "version": "1.0",
        "title": "接口文档",
        "contact": {},
        "license": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
...
    "paths": {
...
        "/v1/bestpractice/check": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "bestPractice",
                "operationId": "bestPracticeUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "name": "objectId",
                    "in": "query",
                    "description": "objectId",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/check/byCluster": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "checkByCluster",
                "operationId": "checkByClusterUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "name": "clusterObjId",
                    "in": "query",
                    "description": "clusterObjId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/records/all": {
            "get": {
                "tags": ["best-practice-controller"],
                "summary": "getBestPracticeRecords",
                "operationId": "getBestPracticeRecordsUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "objectId",
                    "in": "query",
                    "description": "objectId",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "type",
                    "in": "query",
                    "description": "type",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/records/bypage": {
            "get": {
                "tags": ["best-practice-controller"],
                "summary": "getBySettingAndPage",
                "operationId": "getBySettingAndPageUsingGET",
                "produces": ["*/*"],
                "parameters": [{
                    "name": "hostSetting",
                    "in": "query",
                    "description": "hostSetting",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "pageNo",
                    "in": "query",
                    "description": "pageNo",
                    "required": true,
                    "type": "integer",
                    "format": "int32"
                }, {
                    "name": "pageSize",
                    "in": "query",
                    "description": "pageSize",
                    "required": true,
                    "type": "integer",
                    "format": "int32"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/all": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "upAll",
                "operationId": "upAllUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/byCluster": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "upByCluster",
                "operationId": "upByClusterUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/byCluster/{clusterObjectId}": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "upByClusterId",
                "operationId": "upByClusterIdUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "name": "clusterObjectId",
                    "in": "path",
                    "description": "clusterObjectId",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/byhosts": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "upByHosts",
                "operationId": "upByHostsUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "hostObjectIds",
                    "description": "hostObjectIds",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/bylist": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "bylist",
                "operationId": "bylistUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "list",
                    "description": "list",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/BestPracticeUpdateByTypeRequest"
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/update/bytype": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "upByHostSetting",
                "operationId": "upByHostSettingUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "request",
                    "description": "request",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/BestPracticeUpdateByTypeRequest"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/virtual-nic": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "getVirtualNicList",
                "operationId": "getVirtualNicListUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "hostObjectIds",
                    "description": "hostObjectIds",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
        "/v1/bestpractice/virtual-nic/update": {
            "post": {
                "tags": ["best-practice-controller"],
                "summary": "updateVirtualNicList",
                "operationId": "updateVirtualNicListUsingPOST",
                "consumes": ["application/json"],
                "produces": ["*/*"],
                "parameters": [{
                    "in": "body",
                    "name": "beans",
                    "description": "beans",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/UpHostVnicRequestBean"
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ResponseBodyBean"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false
            }
        },
...
    },
    "definitions": {
        "BestPracticeUpdateByTypeRequest": {
            "type": "object",
            "properties": {
                "hostObjectIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "hostSetting": {
                    "type": "string"
                }
            },
            "title": "BestPracticeUpdateByTypeRequest"
        },
...
        "ResponseBodyBean": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                },
                "description": {
                    "type": "string"
                }
            },
            "title": "ResponseBodyBean"
        },
...
        "UpHostVnicRequestBean": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string"
                },
                "hostObjectId": {
                    "type": "string"
                }
            },
            "title": "UpHostVnicRequestBean"
        },
...
    }
}